+++ packages/shopify-app-remix/src/server/future/flags.ts
@@ -1,0 +1,26 @@
+export interface FutureFlags {
+  /**
+   * When enabled, returns the same `admin` context (`AdminApiContext`) from `authenticate.webhook` that is returned from `authenticate.admin`.
+   *
+   * @default false
+   */
+  v3_webhookAdminContext?: boolean;
+
+  /**
+   * When enabled authenticate.public() will not work.  Use authenticate.public.checkout() instead.
+   *
+   * @default false
+   */
+  v3_authenticatePublic?: boolean;
+}
+
+export type FutureFlagOptions = FutureFlags | undefined;
+
+export type FeatureEnabled<
+  Future extends FutureFlagOptions,
+  Flag extends keyof FutureFlags,
+> = Future extends FutureFlags
+  ? Future[Flag] extends true
+    ? true
+    : false
+  : false;