+++ packages/shopify-app-remix/src/server/shopify-app.ts
@@ -1,9 +1,7 @@
 import '@shopify/shopify-api/adapters/web-api';
 import {
   ConfigInterface as ApiConfig,
-  ConfigParams,
   LATEST_API_VERSION,
-  Shopify,
   ShopifyError,
   ShopifyRestResources,
   shopifyApi,
@@ -57,7 +55,7 @@ export function shopifyApp<
   Resources extends ShopifyRestResources,
   Storage extends SessionStorage,
 >(appConfig: Config): ShopifyApp<Config> {
-  const api = deriveApi<Resources>(appConfig);
+  const api = deriveApi(appConfig);
   const config = deriveConfig<Storage>(appConfig, api.config);
   const logger = overrideLogger(api.logger);
 
@@ -77,8 +75,9 @@ export function shopifyApp<
     registerWebhooks: registerWebhooksFactory(params),
     authenticate: {
       admin: oauth.authenticateAdmin.bind(oauth),
-      public: authenticatePublicFactory<Resources>(params),
+      public: authenticatePublicFactory<Config['future'], Resources>(params),
       webhook: authenticateWebhookFactory<
+        Config['future'],
         Resources,
         keyof Config['webhooks'] | MandatoryTopics
       >(params),
@@ -113,9 +112,7 @@ function isSingleMerchantApp<Config extends AppConfigArg>(
   return config.distribution === AppDistribution.SingleMerchant;
 }
 
-function deriveApi<Resources extends ShopifyRestResources>(
-  appConfig: AppConfigArg,
-): Shopify<Resources> {
+function deriveApi(appConfig: AppConfigArg) {
   let appUrl: URL;
   try {
     appUrl = new URL(appConfig.appUrl);
@@ -137,7 +134,7 @@ function deriveApi<Resources extends ShopifyRestResources>(
     userAgentPrefix = `${appConfig.userAgentPrefix} | ${userAgentPrefix}`;
   }
 
-  const cleanApiConfig: ConfigParams = {
+  return shopifyApi({
     ...appConfig,
     hostName: appUrl.host,
     hostScheme: appUrl.protocol.replace(':', '') as 'http' | 'https',
@@ -145,9 +142,7 @@ function deriveApi<Resources extends ShopifyRestResources>(
     isEmbeddedApp: appConfig.isEmbeddedApp ?? true,
     apiVersion: appConfig.apiVersion ?? LATEST_API_VERSION,
     isCustomStoreApp: appConfig.distribution === AppDistribution.ShopifyAdmin,
-  };
-
-  return shopifyApi<Resources>(cleanApiConfig);
+  });
 }
 
 function deriveConfig<Storage extends SessionStorage>(
@@ -170,6 +165,7 @@ function deriveConfig<Storage extends SessionStorage>(
     useOnlineTokens: appConfig.useOnlineTokens ?? false,
     hooks: appConfig.hooks ?? {},
     sessionStorage: appConfig.sessionStorage as Storage,
+    future: appConfig.future ?? {},
     auth: {
       path: authPathPrefix,
       callbackPath: `${authPathPrefix}/callback`,